Work in progress before parking this for a while
I have started trying to teach ApexSOQLInjection to understand variable scopes. This is finding a lot more
that needs to be solved.

The code has the start of recursive analysis of the code. It can track ownership of variables.
It now needs the block analysis filling in. See the comment in that method.
From there it needs to find and check the database calls.

These changes leave things still undone. It cannot track mutation well so marks mutable variables as
unsafe. This is a regression.

The new code needs to use the new RuleContext methods to report violations.