Checkpoint commit to back up work in progress
I have started work on changing the SOQL Injection Rule to understand variables
and their scopes. The code at present is using "roving printf()" debugging
techniques to help me to test ideas and understand the AST and its visitor better.
I have been unable to make a working develop/debug environemt in Eclipse.

Focus is on an Expression Validator. This uses knowledge of Variables to determine
if the result of an Expression is sage for use in a SOQL query. I have found that
the visitor can also visit and identify risky Database calls. A logic conclusion of
this would be to move towards an entirely Visitor based solution, but this will be a
few refactors down the line.

I can identify fields and functions. I can process the functions. At the moment the
code to analyse Fields is pessimistic. If it's mutable it's unsafe (unless intrinsically
a safe type). A later refactor could see the code understanding that a private member
is never mutated or is only ever mutated safely. These are all future ideas that
ideally would have their own Issues. At the moment I'm hoping to define the Rule logically
and maybe it will solve a lot of these naturally.